# perf: optimize scheduler queueJob performance